loads everything in page
